remove winston and break out peer logger to smaller units
