+++ package.json
@@ -65,8 +65,7 @@
     "glob": "10.3.10",
     "hardhat": "^2.22.1",
     "hyperswarm": "4.3.5",
-    "prompt": "1.3.0",
-    "winston": "^3.18.3"
+    "prompt": "1.3.0"
   },
   "engines": {
     "yarn": ">=1.15"
